NodeHost exported
