+++ packages/compiler/test/config/config.ts
@@ -4,8 +4,8 @@ import { fileURLToPath } from "url";
 import { CadlConfigJsonSchema } from "../../config/config-schema.js";
 import { CadlRawConfig, loadCadlConfigForPath } from "../../config/index.js";
 import { createSourceFile } from "../../core/diagnostics.js";
+import { NodeHost } from "../../core/host.js";
 import { SchemaValidator } from "../../core/schema-validator.js";
-import { NodeHost } from "../../core/util.js";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
