+++ packages/compiler/testing/test-host.ts
@@ -3,11 +3,11 @@ import { readFile } from "fs/promises";
 import glob from "glob";
 import { fileURLToPath, pathToFileURL } from "url";
 import { createSourceFile, logDiagnostics, logVerboseTestOutput } from "../core/diagnostics.js";
+import { NodeHost } from "../core/host.js";
 import { CompilerOptions } from "../core/options.js";
 import { getAnyExtensionFromPath, resolvePath } from "../core/path-utils.js";
 import { createProgram, Program } from "../core/program.js";
 import { CompilerHost, Diagnostic, Type } from "../core/types.js";
-import { NodeHost } from "../core/util.js";
 import { expectDiagnosticEmpty } from "./expect.js";
 import { BasicTestRunner, createTestWrapper } from "./test-utils.js";
 import {
